Moved appcenter-node-client back to dependency for now.
