+++ gulpfile.js
@@ -187,13 +187,15 @@ gulp.task("clean", function () {
     var del = require("del");
     var pathsToDelete = [
         "src/**/*.js",
+
         "src/**/*.js.map",
         "test/**/*.js",
         "test/**/*.js.map",
         "out/",
+        "!src/extension/appcenter/api/app-center-node-client/**/*.js",
+        "!src/extension/codepush/api/codepush-sdk/**/*.js",
         "!test/resources/sampleReactNative022Project/**/*.js",
-        ".vscode-test/",
-        "!src/extension/appcenter/lib/**/*.js",
+        ".vscode-test/"
     ]
     return del(pathsToDelete, { force: true });
 });
