+++ src/extension/appcenter/api/appCenterClientCredentials.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-import AppCenterClientCredentials = require("./app-center-node-client/src/appCenterClientCredentials");
+import AppCenterClientCredentials = require("app-center-node-client/src/appCenterClientCredentials");
 export { AppCenterClientCredentials };
