+++ src/extension/appcenter/api/createClient.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-import AppCenterClient from "./app-center-node-client";
+import AppCenterClient from "app-center-node-client";
 import { AppCenterClientCredentials } from "./appCenterClientCredentials";
 import * as Q from "q";
 import { Profile } from "../auth/profile/profile";
