+++ src/extension/appcenter/api/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-import AppCenterClient from "./app-center-node-client";
-import * as models from "./app-center-node-client/models";
+import AppCenterClient from "app-center-node-client";
+import * as models from "app-center-node-client/models";
 
 export { AppCenterClient, models };
 export { AppCenterClientFactory, createAppCenterClient, getQPromisifiedClientResult } from "./createClient";
