+++ src/extension/appcenter/lib/app-center-node-client/.gitignore
@@ -1,0 +1,4 @@
+.DS_Store
+node_modules
+npm-debug.log
+.idea
