+++ src/extension/appcenter/lib/app-center-node-client/src/appCenterClientCredentials.d.ts
