+++ src/extension/appcenter/lib/app-center-node-client/src/codepush/models/index.d.ts
